# api-docs/openAPI-specification.yml
openapi: 3.0.1
info:
  version: "1.0.0"
  title: RUMbus API
  description: >-
    RUMbus is a system to provide real-time information on the location of
    buses, as well as estimated arrival times at stops for the transportation
    service of the University of Puerto Rico at Mayaguez (UPRM).
...
paths:
  /announcements/search:
    get:
      tags:
        - Announcements
      summary: Search for announcements with the specified parameters
      parameters:
        - name: id
          in: query
          schema:
            type: string
          example: Announcement id
        - name: expire_at
          in: query
          schema:
            type: string
          example: Expiration date of the announcement
        - name: headline
          in: query
          schema:
            type: string
          example: Announcement headline
        - name: content
          in: query
          schema:
            type: string
          example: Announcement content
        - name: user_id
          in: query
          schema:
            type: string
          example: ID of the user that created the announcement
        - name: skip
          in: query
          schema:
            type: string
          example: Parameter for pagination
        - name: limit
          in: query
          schema:
            type: string
          example: Parameter for pagination
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    user_id:
                      type: string
                    expire_at:
                      type: string
                    headline:
                      type: string
                    content:
                      type: string
                    created_on:
                      type: string
                    updated_on:
                      type: string
              examples:
                '/announcements/search?headline=Out of Service':
                  value: >-
                    [{"id":"2c584b22-cc58-4f7b-9871-32fe4b7f698b","user_id":"f398fece-a30a-4004-bdd7-1d559db7cbcb","headline":"Out
                    of
                    Service","content":"content","expire_at":"2021-03-21T00:27:10.000Z","updated_on":"2020-07-03T04:00:29.000Z","created_on":"2020-07-03T04:00:29.000Z"}]
        '500':
          $ref: '#/components/responses/500'
...
  /announcements/{announcement_id}:
    put:
      tags:
        - Announcements
      parameters:
        - in: path
          name: announcement_id
          schema:
            type: string
          required: true
          description: ID of the announcement
      summary: Update announcement by its id
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                expire_at:
                  type: string
                headline:
                  type: string
                content:
                  type: string
            examples:
              '0':
                value: |-
                  {
                      "headline": "Updated headline"
                  }
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - Announcements
      parameters:
        - in: path
          name: announcement_id
          schema:
            type: string
          required: true
          description: ID of the announcement
      summary: Delete announcement by its id
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get announcement by its id
      tags:
        - Announcements
      parameters:
        - in: path
          name: announcement_id
          schema:
            type: string
          required: true
          description: ID of the announcement
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  user_id:
                    type: string
                  headline:
                    type: string
                  expire_at:
                    type: string
                  content:
                    type: string
                  updated_on:
                    type: string
                  created_on:
                    type: string
              examples:
                '0':
                  value: >-
                    {"id":"ab6f1863-d53f-44ab-91e0-d741cbbe49d0","user_id":"418e5cd2-074c-4d1d-b53e-6cd6d7267b04","headline":"Test
                    headline","content":"This is the
                    content","expire_at":"2021-01-21T00:27:10.000Z","updated_on":"2020-06-28T22:50:19.000Z","created_on":"2020-06-28T22:50:19.000Z"}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"NotFound","message":"Announcement with id:
                    ab6f1863-d53f-44ab-91e0-d741cbbe49d not found"}
        '500':
          $ref: '#/components/responses/500'
...
  /devices:
    post:
      tags:
        - Devices
      summary: Create device
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
            examples:
              '0':
                value: '{"name": "TestDevice"}'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  device_id:
                    type: string
                  otp:
                    type: string
              examples:
                '0':
                  value: >-
                    {"device_id":"8717ceb1-6976-4095-b8e6-9e0c37561b75","otp":"b75989"}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: '{"code":"BadRequest","message":"Need to provide a name."}'
        '500':
          $ref: '#/components/responses/500'
    get:
      tags:
        - Devices
      summary: Get all devices
      parameters:
        - name: skip
          in: query
          schema:
            type: string
          description: Parameter for pagination
        - name: limit
          in: query
          schema:
            type: string
          description: Parameter for pagination
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    imei:
                      type: string
                    created_on:
                      type: string
                    updated_on:
                      type: string
              examples:
                '0':
                  value: >-
                    [{"id":"8717ceb1-6976-4095-b8e6-9e0c37561b75","imei":null,"name":"TestDevice","updated_on":"2020-03-23T03:24:54.000Z","created_on":"2020-03-23T03:24:54.000Z"}]
                "1":
                  value: []
        '500':
          $ref: '#/components/responses/500'
  /devices/verify:
    post:
      tags:
        - Devices
      summary: Verify device
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [imei, otp]
              properties:
                imei:
                  type: string
                otp:
                  type: string
            examples:
              '0':
                value: |-
                  {
                      "imei": "5ef7cecc=",
                      "otp": "5b84d4"
                  }
      responses:
        '202':
          description: Auto generated using Swagger Inspector
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                '0':
                  value: '{"message":"Device was succesfully verified."}'
        '401':
          description: Auto generated using Swagger Inspector
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                '0':
                  value: '{"message":"The one time password is not valid."}'
        '500':
          $ref: '#/components/responses/500'
  /devices/search:
    get:
      tags:
        - Devices
      summary: Search for devices with the specified parameters
      parameters:
        - name: name
          in: query
          schema:
            type: string
          description: Device name
        - name: id
          in: query
          schema:
            type: string
          description: Device id
        - name: imei
          in: query
          schema:
            type: string
          description: Device imei
        - name: skip
          in: query
          schema:
            type: string
          description: Parameter for pagination
        - name: limit
          in: query
          schema:
            type: string
          description: Parameter for pagination
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    imei:
                      type: string
                    created_on:
                      type: string
                    updated_on:
                      type: string
              examples:
                '0':
                  value: >-
                    [{"id":"8717ceb1-6976-4095-b8e6-9e0c37561b75","imei":"5ef7cecc=","name":"TestDevice","updated_on":"2020-03-23T03:24:54.000Z","created_on":"2020-03-23T03:24:54.000Z"}]
                '1':
                  value: []
        '500':
          $ref: '#/components/responses/500'
  /devices/{device_id}:
    get:
      tags:
        - Devices
      summary: Get device by its id
      parameters:
        - in: path
          name: device_id
          schema:
            type: string
          required: true
          description: ID of the device to fetch
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  imei:
                    type: string
                  created_on:
                    type: string
                  updated_on:
                    type: string
              examples:
                '0':
                  value: >-
                    {"id":"8717ceb1-6976-4095-b8e6-9e0c37561b75","imei":"5ef7cecc=","name":"TestDevice","updated_on":"2020-03-23T03:24:54.000Z","created_on":"2020-03-23T03:24:54.000Z"}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"BadRequest","message":"Need to provide an ID to
                    look for."}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
              examples:
                '0':
                  value: >-
                    {"code":"NotFound","message":"Device with id:
                    5422013c-88a2-4632-9842-0720032bcaa6 not found"}
        '500':
          $ref: '#/components/responses/500'
...
components:
  responses:
    500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
          examples:
            'Server Error':
              value: '{"code":"InternalServer","message":"An internal server error ocurred"}'